Refactor LayerPosition and VisualLineLinkText classes to use partial methods for property registration and link decoration
